move verifier test to separate file
